# Optimise `location_key` in `Prosopite.subscribe`